// apps/frontend/mobile/src/api/generated/mobile-gateway/models/genderEnum.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Blockout Mobile Gateway API
 * OpenAPI spec version: 2.0.0
 */

/**
 * Gender category used by Blockout competition data.
 */
export type GenderEnum = (typeof GenderEnum)[keyof typeof GenderEnum];

export const GenderEnum = {
  M: 'M',
  F: 'F',
  O: 'O',
} as const;